added test for portfolio,qualification,videos and photos

+++ future-authorization-server/src/main/java/com/unionbankng/future/authorizationserver/services/ExperienceService.java
@@ -3,11 +3,16 @@ package com.unionbankng.future.authorizationserver.services;
 import com.unionbankng.future.authorizationserver.entities.Experience;
 import com.unionbankng.future.authorizationserver.pojos.ExperienceRequest;
 import com.unionbankng.future.authorizationserver.repositories.ExperienceRepository;
+import com.unionbankng.future.futureutilityservice.grpcserver.BlobType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.io.IOException;
 import java.util.Optional;
 
 @Service
@@ -15,6 +20,7 @@ import java.util.Optional;
 public class ExperienceService {
 
     private final ExperienceRepository experienceRepository;
+    private final FileStorageService fileStorageService;
 
     public Page<Experience> findAllByUserId (Long userId, Pageable pageable){
         return experienceRepository.findByUserId(userId,pageable);
@@ -29,10 +35,16 @@ public class ExperienceService {
     }
 
     public void deleteById (Long id){
+
+        Experience experience = experienceRepository.findById(id).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND, "Experience Not Found"));
+        int status = fileStorageService.deleteFileFromStorage(experience.getMedia(),BlobType.IMAGE);
+        if(status != 200)
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Something went wrong");
+
         experienceRepository.deleteById(id);
     }
 
-    public Experience saveFromRequest (ExperienceRequest request,Experience experience){
+    public Experience saveFromRequest (MultipartFile file,ExperienceRequest request, Experience experience) throws IOException {
         experience.setUserId(request.getUserId());
         experience.setCompany(request.getCompany());
         experience.setCurrent(request.getCurrent());
@@ -41,7 +53,10 @@ public class ExperienceService {
         experience.setEmploymentType(request.getEmploymentType());
         experience.setEndDate(request.getEndDate());
         experience.setHeadline(request.getHeadline());
-        experience.setMedia(request.getMedia());
+        if (file != null){
+            String source = fileStorageService.storeFile(file, request.getUserId(), BlobType.IMAGE);
+            experience.setMedia(source);
+        }
         experience.setTitle(request.getTitle());
         return experienceRepository.save(experience);
     }
